refactor(recombination): split inline test modules into topic files
- Move the two large inline `mod tests` blocks out of `recombination.rs` and `recombination_estimate.rs` into topic-scoped files under `analyze/__tests__/`, sharing observation builders, tree fixtures, and the brute-force oracles through one helper module instead of duplicating them per file.
- Widen a few internals (`WILDTYPE`, `RECOMBINANT`, `intervals_sorted_disjoint_nonempty`, `log_sum_exp_2`, `log_emission`, `median`, `as_probability`) to `pub(crate)` so the relocated white-box tests reach them; the validation-gated `RecombinationHmmParams` fields and the `RecombinationHmmParamsRaw` wire type stay private, with `new` asserted through its getters.
